Bot Can Be Mentioned Anywhere + More Responses
Changed the way the code detects mentioning the bot so that it can be anywhere in the code. Also seems to break if you include punctuation in the message. Will have to strip punctuation out before testing in the future too.

Responses from line 239-274 are now implemented. (How are you - nice)