Update Conversation.cs
Switched from string.Contains() to Regex.IsMatch() for better matching.

Example, with Contains and checking for "nice" then saying "nice" would trigger it, but so would saying "Denice" which is a little too broad. Using Regex for the recognition is a little slower on calculation, but you can check for word boundaries e.g. the pattern "\bnice\b" so that "nice" only matches if it's a standalone word with a boundary on both sides.

Started laying out the logic for the word recognition, waiting on Destiny to clarify a little bit on if this is right or not.