Update Conversation.cs
Switched to using Regex for a lot of the calculations, which may be slightly slower but will be more precise.

Started adding the framework for the word association bit. Currently can split sentences into arrays, count words, etc. No DB mapping yet, and it treats anyone as a single person, no per-user split.